Update 155_RestApiDocuments.html.md
Did the same thing on this one to remove other small stuff that doesn't make sense anymore. The biggest change is moving the RNs to the bottom. I had it wrong on the outline. Also, want to see what the title looks like as a H2 and removed the metadata.